# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.RecordResponse:
    properties:
      score_count:
        type: integer
      score_time:
        type: integer
    type: object
...
  models.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: lp.pektezol.dev
info:
  contact: {}
  description: Backend API endpoints for Least Portals Hub.
  license:
    name: GNU Affero General Public License, Version 3
    url: https://www.gnu.org/licenses/agpl-3.0.html
  title: Least Portals Hub
  version: "1.0"
paths:
...
  /demos:
    get:
      consumes:
      - application/json
      description: Get demo with specified demo uuid.
      parameters:
      - description: Demo UUID
        in: query
        name: uuid
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Demo File
          schema:
            type: file
      tags:
      - demo
...
  /maps/{mapid}/record:
    post:
      consumes:
      - multipart/form-data
      description: Post record with demo of a specific map.
      parameters:
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Host Demo
        in: formData
        name: host_demo
        required: true
        type: file
      - description: Partner Demo
        in: formData
        name: partner_demo
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.RecordResponse'
              type: object
      tags:
      - maps / leaderboards
  /maps/{mapid}/record/{recordid}:
    delete:
      description: Delete record with specified map and record id.
      parameters:
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      - description: Record ID
        in: path
        name: recordid
        required: true
        type: integer
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - maps / leaderboards
...
swagger: "2.0"